Verify tests
