+++ sdk/python/tests/integration/scaffolding/test_partial_apply.py
@@ -5,7 +5,7 @@ import pytest
 from feast import BigQuerySource, Entity, FeatureView, Field
 from feast.types import Float32, String
 from tests.utils.cli_utils import CliRunner, get_example_repo
-from tests.utils.online_read_write_test import basic_rw_test
+from tests.utils.online_read_write_test_utils import basic_rw_test
 
 
 @pytest.mark.integration
